Update main.py
updates for kicks and giggles.
When the PySimpleGUI window is closed, the read() method returns a tuple with two elements: the event that was triggered (in this case, either "OK" or "Cancel") and a dictionary with the input values

The read() method replaced with (). --  read() is no longer a valid method for creating a PySimpleGUI window.

The email and password variables were being overwritten with Selenium web elements.